feat: Rename items to tags [BREAKING CHANGE]
